\name{save_config}
\alias{save_config}
\title{
  Save configuration file
}
\description{
This is a function to save information of the configuration file. When users change the content of the configuration. They must call this function to perform the tasks for saving. 
}
\usage{
  save_config(data = "")
}
\arguments{
  \item{data}{contents of the configuration file}
}
\keyword{save}
\keyword{config}
