\name{trim}
\alias{trim}
\title{
  Remove whitespace from both sides of a string
}
\description{
  Remove all spaces from text except for single spaces between words
}
\usage{
    trim(x)
}
\arguments{
  \item{x}{ is a string that we want to delete whitespace from both sides}
}
\examples{
  str = " Hello World! "
  trim(str)
}
\keyword{trim}
