\name{xentity}
\alias{xentity}
\alias{xrelation}
\title{
List of entities or relations
}
\description{
  Show all entities or relations
}
\usage{
  xentity()
  xrelation()
}
\value{
  \item{list}{return a list of entities or relations}
}
\seealso{
  \code{\link{xshow}} display results
}
\examples{
  xentity()
  xrelation()
}
\keyword{list entities}
\keyword{list relations}